+++ src/pages/nodejs.tsx
@@ -1,5 +1,5 @@
 import { FC } from 'react';
-import { About } from '@/widgets/about';
+import { AboutCourse } from '@/widgets/about-course';
 import { Breadcrumbs } from '@/widgets/breadcrumbs';
 import { Certification } from '@/widgets/certification';
 import { Communication } from '@/widgets/communication';
@@ -17,7 +17,7 @@ export const Nodejs: FC = () => {
       <CourseMain courseName="Node.js" />
       <Breadcrumbs />
       <TrainingProgram courseName={COURSE_NAME} />
-      <About courseName={COURSE_NAME} />
+      <AboutCourse courseName={COURSE_NAME} />
       <Certification courseName={COURSE_NAME} />
       <Communication courseName={COURSE_NAME} />
       <Required courseName="nodejs" />
