+++ src/widgets/about-course/index.ts
@@ -1,0 +1,1 @@
+export { AboutCourse } from './ui/about-course';
