+++ src/widgets/about-course/ui/about-course-grid/about-course-grid.module.scss
