+++ src/widgets/about-course/ui/about-course-grid/about-course-grid.test.tsx
@@ -1,6 +1,6 @@
 import { render, screen } from '@testing-library/react';
 import { describe, expect, it } from 'vitest';
-import { InfoGrid } from './info-grid';
+import { InfoGrid } from './about-course-grid';
 import { MOCKED_IMAGE_PATH } from '@/shared/__tests__/constants';
 
 describe('InfoGrid component', () => {
