+++ src/widgets/about-course/ui/about-course.module.scss
